Implemented involuntaryAuthorizationDecrease for WalletRegistry
involuntaryAuthorizationDecrease is used by T staking contract to inform the
application the authorization has been decreased for the given staking provider
involuntarily, as a result of slashing.

If the operator is not known (`registerOperator` was not called),
the function does nothing. The operator was never in a sortition pool so there
is nothing to update.

If the operator is known, sortition pool is unlocked, and the operator is in the
sortition pool, the sortition pool state is updated. If the sortition pool is
locked, update needs to be postponed. Every other staker is incentivized to call
`updateOperatorStatus` for the problematic operator to increase their own
rewards in the pool.