Store uncompressed public key
We want to store a public key of the wallet so it can be used by wallet
owner (Bridge) for transactions handling. The Bridge will use both
compressed and uncompressed public keys. We store the longer one
(uncompressed) that can be eaisly converted to the compressed format.

We also renamed publicKeyHash to walletID as this reflects better what
this property is. Using a name `publicKeyHash` was also confusing in
external owner contract as there may be another key hashes in use, but
this is an implementation detail what hashing function we use.