Use the same governance delay for all parameters
2-week critical governance parameter does not make sense for ECDSA WalletRegistry
given that wallets will have to remain active for months. We could swallow it for
the RandomBeacon but for WalletRegistry 2 weeks does not make any sense and 2
months is not enough.

One option would be to eliminate governance contracts completely and just deploy
TimeLock contract. On the other hand, TimeLock updates are more complex to set
up than simple calls to the existing governance contracts, and the structure of
update* functions in RandomBeacon and WalletRegistry is more complicated.

Here we introduce just one governable delay for parameter updates and eliminate
the separation between critical/standard parameters given it creates a false
assumption of safety.