Add workflow publishing ecdsa contracts for dev env
We're adding a workflow which will publish a package with compiled
`ecdsa` contracts. Package will be meand for development purposes.

The `-dev` suffix will be used in the name of the published package.
The workflow will be triggered every time a change to contracts or
dependencies gets pushed to `main`. Also a change to the workflow file
will trigger the workflow.

Note: We're temporarily publishing to NPM registry using `--dry-run`
flag. The flag will be removed before merging of the changes
to the `main` branch. Some other temporary configuration is used as
well, all will be removed before merge to `main`.