PlainTransferEscrowBeneficiary implementation
PlainTransferEscrowBeneficiary is a beneficiary contract that can
receive a withdrawal order from PhasedEscrow and BatchedPhasedEscrow
contracts. It immediately sends the received tokens to the recipient
set by the beneficiary deployer.
The recipient does not have to implement any custom functions, it can
be a smart contract or EOA - beneficiary is executing a plain ERC20
transfer.