Improved test coverage for remainingAuthorizationDecreaseDelay
Most of the corner cases is tested as a part of tests of other
functions, such as the corner case when operator is not registered for
the staking provider. Tests added make sure the most basic path works
as expected.

I have also removed instruction increasing time in one of eligibleStake
tests (added in the previous commit). I have realized this line is not
needed because operator is not registered for the staking provider and
authorization decrease request can be approved immediately.