Adjusting gas for approveDkgResult
It is tricky to calcualte the exact gas usage for approveDkgResult
function. Especially if the refund covers two call: submit & approval of
DKG result. The submision part is fluctuates around 300k. In approal
part the fluctuation is greater hence the gas is calcualted
programmatically of the fly. The gas numbers and precisions were
selected for the gas prices between 10-200gwei.