Replaced Random Beacon failed heartbeat with operator inactivity claim
The goal of the inactivity claim is to have the group punish operators
who are constantly inactive and not contributing to the work by
excluding them from sortition pool rewards.

The goal of the wallet heartbeat failure - present only in ECDSA - is to
transfer funds to another wallet from the one who failed the heartbeat.

Inactivity claim in ECDSA will be only about excluding operators from
rewards and will not require moving funds.

For this reason, we need to keep these two actions separate and properly
named:
- inactivity claim: exclude operators from rewards
- heartbeat failure: move funds to another safe wallet (just ECDSA)