Add workflow publishing random-beacon contracts for dev env
We're adding a workflow which will publish a package with compiled
`random-beacon` contracts. Package will be meant for development
purposes.

The `-dev` suffix will be used in the name of the published package.
The workflow will be triggered every time a change to contracts or
dependencies gets pushed to `main`. Also a change to the workflow file
will trigger the workflow.