Use `development` tag to configure dependency between packages
Some time ago we temporarily changed `@keep-network/random-beacon`
dependency in `solidity/ecdsa/package.json` from
`>2.0.0-dev <2.0.0-ropsten` to `../random-beacon`. This was only a
tempoorary change, to allow for passing of checks in PR2856. Now we want
to set the dependency to again point to the latest package meant for the
development environment. We discovered, that using
`>2.0.0-dev <2.0.0-ropsten` range to do that is not ideal (if one of the
packages in the range is tagged 'latest', but there are newer packages
which seem to fit the range, those newer packages won't be recognize as
fitting the range). As an alternative, we propose to refer to the latest
package by using following dependency configuration:
```
"@keep-network/random-beacon": "development"
```
This will always point us to the package tagged `deployment`. What we
also want to make sure, is to publish under this tag only the packages
which were built on the `main` branch.