notifyOperatorInactivity for WalletRegistry
Function lets the wallet signing group operators notify about operators who are
inactive in the group. Using this function, a majority of the wallet signing
group can decide about punishing specific group members who constantly fail
doing their job. If the provided claim is proved to be valid and signed by
sufficient number of group members, operators of members deemed as inactive are
banned for sortition pool rewards for duration specified by
`sortitionPoolRewardsBanDuration` parameter. The sender of the claim must be one
of the claim signers. This function can be called only for registered wallets.