Implemented WalletRegistry.closeWallet
So far this function was just a stub and now the implementation is there. The
logic is pretty simple - we emit an event and remove the wallet from the
registry.

The idea behind removing the wallet from the registry is that we do not want to
allow reporting inactivity for a closed wallet and at the same time, we do not
want to maintain separate wallet states in `WalletRegistry` - it's happening
already in tBTCv2 `Bridge` contract.