Added unit test making sure nonces are not deleted for closed wallet
In case a new wallet with the same ID (public key) is registered, we
need to make sure old inactivity claims do not work. Wallet owner may decide to
revert in `__ecdsaWalletCreatedCallback` in case it is important to have the
public keys unique but from the perspective of `WalletRegistry` this is not
strictly required as long as inactivity claim nonces are not cleared up upon
closing the wallet.