Use unique addresses for staking roles
To have the tests that are more reliable we neeed to ensure unique
address is used for each of the roles registered in Token Staking.
We need an unique address for 64 group members for each role: owner,
staking provider, operator, beneficiary and authorizer.

We introduced `unnamedSignersOffset` to specify from which unnamed
signer account we should start obtaining signers. This is for a case
where we want to pull some other accounts e.g. for third party and
don't want to overlap.