Use allowUnlimitedContractSize for hardhat
Our contracts grew to the point where a RandomBeaconStub is problematic to deploy.
The RandomBeacon size is fine though.

We can safely ignore contract size on deployment in hardhat network as
contract size is verified by `contractSizer` and it will alert us if any
of the contracts is too big.