Check there is enough signers configured
We want to have unique signers for operators registration for each of 5
roles. This requires hardhat network configuration account count to be
aligned. We want to ensure the configuration is correct and output a
meaningfull error if it's not. Without that the test execution fails
with an error that doesn't give much clue.