Use unique accounts for operators staking and registration
We need to make our tests more roboust by using unique addresses for
each role that is related to staking and operator registration. This
requires to configure hardhat network with more than 500 accounts so
each address is unique and don't overlap.
Thanks to that the tests are more reliable and reduce possibility to
give a false positive results.