Remove Ownable from Reimbursable
Reimbursable is an abstract contract with a function
`updateReimbursementPool` that should be able to be executed only by a
specific role. Using Ownable for this contract was not necessary,
especially that this causes problems when using upgradable
implementations, where we should use OwnableUpgradable instead of
Ownable.
With `onlyReimbursableAdmin` virtual modifier we leave it to the
implementation to check the role according to their needs.