Tweaking the gas offset and precision after the code changes
The gas offset for the DKG approval is lowered to hit most of the upper
limits of the refund. There's only 1 test where the precision has
changed from 0.002 ETH to 0.0023. This is the trade-off between all the
tests and the gas offset.

The gas offset for the relay entry submission can be lowered after the
code change.