Check wallet membership
This change adds `isWalletMember` function
to the ECDSA wallet registry. That function
allows checking whether the given operator
is a member of the given wallet signing group.
In order to make a successful check, the tested
operator address must actually belong to a
sortition pool operator and the passed wallet
members identifiers must be equal to the hashed
identifiers held by the contract.