Update table with parameters
Since last version of the table was commiteed, a bunch of changes in
contracts has occured. Some of the parameters were removed and some were
addet. This commit updates the list to reflect all the changes. Also
some small changes relating formatting, interpunction, etc were applied.

Parameters that were removed from the list:
- `offchainDkgTime`
- `relayRequestFee`
- `dkgResultSubmissionReward`
- `sortitionPoolUnlockingReward`
- `ineligibleOperatorNotifierReward`

Parameters that were added to the list:
- `singnatureByteSize`
- `_owner`
- `dkgResultSubmissionGas`
- `dkgResultApprovalGasOffset`
- `notifyOperatorInactivityGasOffset`
- `relayEntrySubmissionGasOffset`
- `authorizedRequesters`
- `governanceDelay`