WalletRegistry.seize function implemented
Allows the wallet owner to add all signing group members of the wallet with the
given ID to the slashing queue of the staking contract. The notifier will
receive reward per each group member from the staking contract notifiers
treasury. The reward is scaled by the `rewardMultiplier` provided as a parameter.

This function will be used by tBTC Bridge contract to slash wallet signing
group members.