Fix the tagging of `random-beacon` and `ecdsa` packages
Our intention is to tag the published NPM packages with `development`
tag when changes to the contract and configuration related files get
modified and those changes land on `main` branch. The configuration of
the steps in the `npm-....yml` workflows responsible for that was wrong.
The format of the `github.ref` variable for the `main` branch is
`refs/heads/main`, not `main`.