Governable contract copied from ecdsa to random-beacon
This contract was originally implemented by @nkuba for ECDSA
`WalletRegistry` but we can use it also for `RandomBeacon`. First, for
consistency, second to decrease its size given we may need space for one
more function and event related to rewards.

Files are copied and not moved so that we can merge into `main`, have a
new version of `@keep-network/random-beacon` pushed to NPM registry and
in a separate PR, import these contracts from `ecdsa` module and remove
them from there.