Expand NPM workflow triggers
We want to publish the new NPM package tagged with `development` every
time changes affecting contracts get merged to `main`. We already
trigger the NPM workflows when changes are applied in couple of files,
but we missed the hardhat configuration in `hardhat.config.ts` which
can also affect the content of the exported package with the
contracts.