Add gaps to structs
We added gaps to structs from the same reason we did it for base
contracts in 1b67c56bbdfe0b9d957e21919de2e16e2499be94.

A structs holds a structure of data that is stored in the storage. In
case we upgrade the contract we may want to add more fields to the
struct, this could affect the storage by shifting other data. So we
added gaps to the structs that hold data that are saved in storage.

We did not add the gaps to the structs that are used as calldata as
those are used just to organize a function arguments, and are not saved
in the storage.

Gaps for structs are calculated so the whole structure has 20 entries.

Ref: https://docs.openzeppelin.com/contracts/4.x/upgradeable#storage_gaps