Ignore contract size on hardhat network
For tests we deploy stub contracts e.g. WalletRegistryStub and it
exceeds the contract size limit, but WalletRegistry is fine. We don't
expect to use the stub contract any further than on hardhat local test
network so we ignore the limit.
We use `hardhat-contract-sizer` plugin to monitor contract size on
compilation and it will let us know once `WalletRegistry` is too big.