Updated initial RandomBeacon parameters
All parameters set in the constructor are initial ones, used at the moment
contracts were deployed for the first time. Parameters are governable and values
assigned in the constructor do not need to reflect the current ones.

Minimum authorization is 40k T.

Authorization decrease delay is 45 days.

Authorization decrease change period is 45 days. It means pending authorization
decrease can be overwriten all the time.

Malicious DKG result slashing amount is set initially to 1% of the minimum
authorization (400 T). This values needs to be increased significantly once the
system is fully launched.

Unauthorized signing slashing amount is set initially to 1% of the minimum
authorization (400 T). This values needs to be increased significantly once the
system is fully launched.

Slashing amount for not providing relay entry on time is set initially to 1% of
the minimum authorization (400 T). This values needs to be increased
significantly once the system is fully launched.

Inactive operators are set as ineligible for rewards for 2 weeks.

Notifier of a malicious DKG result receives 100% of the notifier reward from the
staking contract.

Notifier of unauthorized signing receives 100% of the notifier reward from the
staking contract.

Notifier of relay entry timeout receives 100% of the notifier reward from the
staking contract.

Inactive operators are set as ineligible for rewards for 2 weeks.

DKG result challenge period length is set to 48h, assuming 15s block time.

DKG result submission timeout, gives each member 20 blocks to submit the result.
Assuming 15s block time, it is ~8h to submit the result in the pessimistic case.

The original DKG result submitter has 20 blocks to approve it before anyone else
can do that.

With these parameters, the happy path takes no more than 56 hours. In practice,
it should take about 48 hours (just the challenge time).

Relay entry soft timeot gives each of 64 members 20 blocks to submit the result.

Relay entry hard timeout is set to ~48h assuming 15s block time.

Callback gas limit is set to 56k units of gas. As of April 2022, it is enough to
store new entry and block number on-chain. If the cost of EVM opcodes change
over time, these parameters will have to be updated.

Group lifetime is set to 45 days assuming 15s block time.

New group is created every 2 relay requests.

This way, even if ECDSA WalletRegistry is the only consumer of the beacon
initially, and relay request is executed every week, we should have 2 active
groups in the system all the time.

Gas parameters were adjusted based on Ethereum state in April 2022. If the cost
of EVM opcodes change over time, these parameters will have to be updated.