Add offsets to refunds for timeout notification transaction
As the WalletRegistry contract is deployed with an upgradable proxy
pattern the costs of transaction execution went up.

For reimbursments we need to align the gas calculation for refund to
cover the higher gas cost.

We cannot modify staticGas parameter that is used by default and was
calculated for standard contract transactions and works fine for the
RandomBeacon contract. We need to add some offset to all the
transactions refunded for WalletRegistry contract.

We need to update the way gas for refund is calculated in
notifySeedTimeout and notifyDkgTimeout transactions, hence we add
offsets in those refunds.

Each transaction has inidividual offset as the gas consumption differs
in them.

notifyDkgTimeout is a special case, where we implemented a concept of
negative gas offset, which results in the offset subtraction. This is
caused by the fact that on a stage where DKG timeout is notified the
data that were stored on previous stages of the DKG are wiped. Those
`delete` operations are recovering some gas for the transaction.