Use _disableInitializers in contructor
We call `_disableInitializers` to lock an implementation contract. We
need to lock context of the implementation so initialize is not called
on this contract.

Note the difference of proxy and implementation contexts. We deploy
implementation and leave it be without initializing it, as the data
won't be stored there anyways. The initializer will be called on proxy
context to initialize data stored in the proxy context.