Move v1 documentation to `./docs-v1` folder
We want to store v1 and v2 documentation in separate places. In this
commit we rename `./docs` to `.docs-v1`. We also adjust the `docs.yml`
GitHub Actions workflow to operate on the correct files. In a separate
commit we'll create a new `./docs` folder which will be intended for
storing v2 documentation.