Workaround bug #86 in hardhat-gas-reporter
There is a bug in hardhat-gas-reporter that we need to workaround in
order to obtain gas report for tests: https://github.com/cgewecke/hardhat-gas-reporter/issues/86
Long story short: we cannot use `hre.ethers` in the deployment script.

Here we introduce an alternative deployment path just to get the report
generated correctly. The alternative path usage is determined by
`GAS_REPORTER_BUG_WORKAROUND` env property set to `true`.

To run the tests with the alternative path to get the gas report execute
`yarn test:gas-reporter-workaround`. It will execute the same test (upgrade
tests were skipped as they are strictly related to the original
deployment path).

`yarn test` is still a preferred way to run the tests, as it does the
deployment exactly the way we want it to test for mainnet. Use the
alternative path only if you wish to get the gas report.

Once a bug in the `hardhat-gas-reporter` is fixed, we should revert all
the changes from this commit.