Export WalletRegistryStub in the npm package
We don't export contracts from `contracts/test` directory, as those are
just stubs that are internal to the project. The exception is
WalletRegistryStub that we use in the deployment script on hardhat
network. We need to include this contract in the package, so the other
projects can use deployment scripts of WalletRegistry as a dependency.