Workaround hardhat-gas-reporter bug with removing --deploy-fixture
In https://github.com/keep-network/keep-core/pull/2970 we introduced a
workaround for a bug in hardhat-gas-reporter plugin
https://github.com/cgewecke/hardhat-gas-reporter/issues/86.

It turned out that there is a much simpler workaround, which is based on
just removing `--deploy-fixture` flag. With the flag removed it is
required that we load a global snapshot of the deployment
(`deployments.fixture()`) instead of loading a specific tag
(`deployments.fixture(<tag>)`. Which we do anyway after one of a
recent code updates.

We removed all the traces of the previous workaround.