Compile external @threshold-network TokenStaking
Define solidity compiler configuration for TokenStaking contract that
reflect the configuration the contract is compiled in the original
repository @threshold-network/solidity-contracts.
With this change the size of the contract for deployment will be the
same as desired in @threshold-network/solidity-contract deployment and
it won't fail deployment due to the contract size.