Use range in the t-network solidity-contracts dependency
We tried using `development` tag to always point to the latest
development package, but that was causing `An unexpected error occurred:
"expected manifest".` error when running `yarn upgrade
@threshold-network/solidity-contracts`. So we're going back to using the
version range. We may use `development` tag in the future, once we
figure out the reason behind the error that we were getting.

It is possible that the error was related to this issue:
https://github.com/yarnpkg/yarn/issues/4731.