Add TEST_USE_STUBS flag
The TEST_USE_STUBS flag was added for usage in tests to determine if a
stub contract should be deployed. Having check based just on the network
type is not enough, as we want to use deployment scripts in other
projects that would use a stub instead of the proper implementation.
If we are executing unit tests that require functions from the stub we
add the flag.