Config git to use `https://` instead of `git://`
The `git://` protocol is no longer supported by Github. This means that in
certain situations installation of the package or update of its dependencies
using NPM may result in errors like `fatal: unable to connect to github.com`.
That is what happenning after executing the Dockerfile and installing some
dependencies of the token dashboard.

As a workaround, we're adding step to the Dockerfile, which configures Git to
use `https://` protocol instead of `git://` when downloading files.