Create a common Makefile for contracts bindings generation
The common Makefile is based on the Makefile used for go bindings
generation for contracts from V1.

It is a common file that should be included in other Makefiles for
projects.

The main difference is that solc generation was extracted to a function
that expects dependencies imports mapping and source file as arguments.