Validating a new group lifetime > 0
Current validation is dependent not only on the authorization decrease
delay but also on the avg. block time. Block time can vary and might
create issues if not properly adjusted. Instead we can only validate if
the lifetime is > 0 and inform the governance in the docs to set a new
value with caution.