Redefine the operator key
So far, we had a mess regarding key types used
in the keep network client implementation.
All keys were actually `ecdsa`-package keys
with the same curve type but different curve
implementations:
- Chain keys used go-ethereum curve implementation
- Network keys used btcec curve implementation
- Operator used an unspecified curve
implementation, depending on the use case

Conversions between those types were crucial
but very error-prone. To improve the situation,
we decided to:
- Redefine the operator key to be agnostic
  regarding curve implementation. It just
  holds the curve name instead.
- Remove the concept of the abstract network keys,
  in favor of specific network key implementations.
  In our case, we now have just the `libp2p keys`
  that use the libp2p-specific curve
  implementation. The `libp2p` package provides
  functions that allow `libp2p keys <-> operator keys`
  conversions
- Remove the concept of the abstract chain keys,
  in favor of specific chain key implementations.
  We now have just the `ethereum key` that use
  go-ethereum-specific curve implementation.
  The `ethereum` package provides functions that
  allow `ethereum keys <-> operator keys`
  conversions.

This changeset breaks the compilation.
The code impacted by the aforementioned changes
will be adjusted in the subsequent commits.