Expose Ethereum Chain implementation
We export Ethereum Chain handle to use in other packages as we are
aiming for packages structure refactoring and we expect to use the
common chain handle in multiple modules.

We also renamed the struct to `Chain` as it's defined in `ethereum`
pacakge and will be used externally as `ethereum.Chain`, which is better
than `ethereum.EthereumChain`.