Adjust the `start` command
So far the ethereum key was converted to the
operator key and then to the network key.
This change simplifies that by converting
the ethereum chain key to the operator key
using the new `ethereum.ChainKeyToOperatorKeyPair`
converter function and passes that key
directly to the `libp2p.Connect` function.
Conversion to the libp2p network key is now
handled inside the `libp2p` package