Beacon: Improve validation of new group's public key
Instead of checking the uniqueness in `approveDkgResult`, when the non-unique
group passed the challenge period, we should check it in `submitDkgResult`. This
way, in an extremely unlikely case, the same group is getting chosen and they
produce the same public key, DKG will not be blocked forever.

The same fix has been applied to ECDSA `WalletRegistry`.

This change increases the gas cost of `submitDkgResult` by ~2640 gas units on
average.

Before/after:
| submitDkgResult   ·  232349  ·  239837  ·  235096  │
| submitDkgResult   ·  234996  ·  242484  ·  237743  │
| approveDkgResult  ·  339712  ·  418859  ·  363560  │
| approveDkgResult  ·  339473  ·  418620  ·  363321  |