Download V2 artifacts in docker build
We download the artifacts for V2 contracts in the docker image building
process.

There is an optional argument `CONTRACTS_NPM_PACKAGE_TAG` that can be
set to a tag with which npm packages should be downloaded (e.g. ropsten,
mainnet). By default if the argument is not set the `development`
artifacts will be downloaded.