Download artifacts as part of go generate
Having a separate bash script for getting the artifacts is another
complication. To simplify the generation we include the download
artifact in the go generate command execution.
It makes sense to pull the artifacts each time the go generate is run as
we want to build the fresh bindings in case the packages were updated.