Use stubs only if explicit flag is set
We want to use stubs in unit tests only. The stub for RandomBeacon
exceeds the gas limit so we need to set `allowUnlimitedContractSize`.
With an env variable `TEST_USE_STUBS_BEACON` we will explicitly
determine that stubs should be used for tests. This has an effect on
other projects using the beacon's deployents scripts as the external
projects won't use stubs.