Use external contracts deployment only in tests
For deployments we expect that external contracts are already deployed
and deployment artifacts are provided. Only in tests we allow the
deployment process to execute also deployment of missing contracts from
their projects. This is why we define `USE_EXTERNAL_DEPLOY` flag.
If the flag is set the scripts will try to deploy external contracts to
use in this project.