Accept single instance of Ethereum chain when creating application chains
We need to ensure there is just one instance of transactionMutex. This mutex is
held by the base Ethereum chain so there has to be only one instance of Ethereum
base chain. The idea is to keep the base chain as a singleton and create
application chains using public functions of the base chain (to be implemented
separately).