Global Connect function to create Beacon and TBTC Ethereum chains
We need to ensure there is just one instance of transaction mutex. This mutex is
held by the base Ethereum chain so there has to be only one instance of Ethereum
base chain. The idea is to call global `Connect` only one time and create all
the chain instances only one time. Creating chain instances outside of `Connect`
is not possible from the outside of `ethereum` package.