Switch `golint` CI job to `gofmt`
Current `golint` CI job raised a lot of
errors in packages that will be removed
soon. In the same time `golint` does not allow
for ignoring those errors and became deprecated.
We decided to remove it in favor of `gofmt`
to maintain formatting rules. We will add
a new linting tool soon.