Mechanism for monitoring sortition pool status
MonitorPool periodically checks the status of the operator in the sortition
pool. If the operator is supposed to be in the sortition pool but is not
there yet, the function attempts to add the operator to the pool. If the
operator is already in the pool and its status is no longer up to date, the
function attempts to update the operator's status in the pool.