Integrated sortition pool monitoring loop with beacon node
The client monitors the sortition pool status now. There is a complication with
ethereum chain - the v2 implementation does not support v1 ethereum chain
functions and the v1 implementation does not support v2 functions. For now, both
instances are passed. In the future, we will decide if we want to port features
to v2 gradually or cut the entire v1 chain in one move altogether.

Note that from now on, `RandomBeacon` contract address must be specified in the
config file.

Old contract addresses still need to be there but the values do not matter.