Move v2 contract bindings to the `pkg/chain/ethereum` package
New v2 contract bindings are Ethereum-specific
so, we move them to the `pkg/chain/ethereum`
package. This way we achieve the desired
chain package structure we agreed on recently.
By the way, some bindings packages were
renamed to align with the naming convention
used in the `pkg` root package and make those
names shorter:
- `random-beacon` -> `beacon`
- `tbtc-v2` -> `tbtc`
- `threshold-network` -> `threshold`