Extract contract addresses from artifacts
We already download artifacts of deployed contracts from NPM registry.
Here we add support of extracting addresses of contracts from the
artifacts. The address is extracted and stored in `_address/<CONTRACT>`
file and then embeded into the go code.

This solution is intended to be used for clients building for specific
environment. For example if we set `CONTRACTS_NPM_PACKAGE_TAG`
environment variable to `goerli`. It will download a package of
contracts deployed on Goerli network and extract addresses from it. The
client will be build specific for Goerli network, without a need of
configuring the addresses in a config file.

There will be a possibility to overwrite the addresses with a
configuration file or environment variables.

Addresses extraction will happen on `go generate ./...` command
execution along with the go bindings generation.

To test the solution for Goerli please run:
`CONTRACTS_NPM_PACKAGE_TAG=goerli go generate./...`

Please note that for `development` which is a default tag the addresses
will be replaced with zeros, as the contracts artifacts that are published with
`development` tags are intended only for ABIs generation, as the
contracts are deployed locally in CI, and won't be accessible anywhere.